change type of srs_toa_ns

+++ openair1/SCHED_NR/phy_procedures_nr_gNB.c
@@ -983,7 +983,7 @@ int phy_procedures_gNB_uespec_RX(PHY_VARS_gNB *gNB, int frame_rx, int slot_rx, N
 
         uint8_t N_ap = 1 << srs_pdu->num_ant_ports;
         uint8_t N_ant_rx = gNB->frame_parms.nb_antennas_rx;
-        int32_t srs_toa_ns[N_ant_rx];
+        int16_t srs_toa_ns[N_ant_rx];
 
         start_meas(&gNB->srs_timing_advance_stats);
         srs_indication->timing_advance_offset =
@@ -999,7 +999,6 @@ int phy_procedures_gNB_uespec_RX(PHY_VARS_gNB *gNB, int frame_rx, int slot_rx, N
           T_INT(slot_rx),
           T_BUFFER(srs_toa_ns, N_ant_rx * sizeof(int32_t)));
 
-        //(int16_t)((((int32_t)srs_indication->timing_advance_offset - 31) * ((int32_t)TC_NSEC_x32768)) >> 15) : 0xFFFF;
         switch (srs_pdu->srs_parameters_v4.usage) {
           case 0:
             LOG_W(NR_PHY, "SRS report was not requested by MAC\n");
